feat&refactor: add executor and redesign storage.

+++ crawler/storage.py
@@ -4,8 +4,11 @@ import os.path
 import shutil
 
 
-class SPECDirectoryStruct:
-    __SPEC = "SPEC"
+def remove_catalog_suffix(file_name):
+    return file_name.rstrip(".csv")
+
+
+class GroupDirectoryStruct:
     __CATALOG = "catalog"
     __RESULTS = "results"
     __BENCHMARKS = "benchmarks.txt"
@@ -15,7 +18,7 @@ class SPECDirectoryStruct:
         assert os.path.exists(os.path.exists(data_dir))
         data_dir = os.path.abspath(data_dir)
         self.group = group
-        self.data_root_path = os.path.join(data_dir, self.__SPEC, group)
+        self.data_root_path = os.path.join(data_dir, group)
         self.benchmarks = []
 
     @property
@@ -36,12 +39,9 @@ class SPECDirectoryStruct:
     def get_results_path(self, benchmark):
         return os.path.join(self.results_path, benchmark)
 
-    def remove_catalog_suffix(self, file_name):
-        return file_name.rstrip(".csv")
-
     @staticmethod
-    def load_exists(data_dir, group="OSG"):
-        new_ds = SPECDirectoryStruct(data_dir, group)
+    def try_load_existed(data_dir, group="OSG"):
+        new_ds = GroupDirectoryStruct(data_dir, group)
         if not os.path.exists(new_ds.data_root_path):
             raise FileNotFoundError("No struct dir exists")
         benchmarks_path = new_ds.benchmarks_file_path
@@ -51,9 +51,14 @@ class SPECDirectoryStruct:
         catalog_path = new_ds.catalog_path
         results_path = new_ds.results_path
         if not os.path.exists(results_path) or not os.path.exists(catalog_path):
-            raise FileNotFoundError("Cannot launch exists, broken structure.")
-        new_ds.verify_benchmarks()
-        new_ds.verify_catalogs()
+            raise FileNotFoundError("Broken structure.")
+        try:
+            new_ds.verify_catalogs()
+            new_ds.verify_results()
+        except FileNotFoundError as e:
+            raise e
+        except NotADirectoryError as e:
+            raise e
 
         return new_ds
 
@@ -74,7 +79,7 @@ class SPECDirectoryStruct:
                 b = row.strip().split("|")[0]
                 self.benchmarks.append(b)
 
-    def build_results(self, benchmark):
+    def rebuild_results(self, benchmark):
         catalog_file_path = self.get_catalog_file_path(benchmark)
         if os.path.exists(catalog_file_path):
             assert os.path.isdir(catalog_file_path)
@@ -89,33 +94,41 @@ class SPECDirectoryStruct:
                 file_type_path = os.path.join(catalog_file_path, field)
                 os.mkdir(file_type_path)
 
-    def verify_benchmarks(self):
+    def verify_catalogs(self):
         assert os.path.isdir(self.catalog_path)
         catalogs = os.listdir(self.catalog_path)
         for elem in catalogs:
-            benchmark = self.remove_catalog_suffix(elem)
+            benchmark = remove_catalog_suffix(elem)
             if benchmark not in self.benchmarks:
                 file_removing = self.get_catalog_file_path(benchmark)
-                assert os.path.isfile(file_removing)
+                if not os.path.isfile(file_removing):
+                    raise FileNotFoundError(f"{file_removing} is not a file.")
                 os.remove(file_removing)
 
-    def verify_catalogs(self):
+    def verify_results(self):
         assert os.path.isdir(self.catalog_path)
         assert os.path.isdir(self.results_path)
         catalogs = os.listdir(self.catalog_path)
-        catalogs = list(map(self.remove_catalog_suffix, catalogs))
+        catalogs = list(map(remove_catalog_suffix, catalogs))
         results = os.listdir(self.results_path)
 
         for elem in results:
             if elem not in catalogs:
                 path_removing = self.get_results_path(elem)
-                assert os.path.isdir(path_removing)
+                if not os.path.isdir(path_removing):
+                    raise NotADirectoryError(f"{path_removing} is not a dir.")
                 shutil.rmtree(path_removing)
 
     def __repr__(self):
         return f"<data_root_path: {self.data_root_path}, benchmarks: {self.benchmarks}>"
 
 
-if __name__ == '__main__':
-    ds = SPECDirectoryStruct("/home/uw2/data")
-    ds.build_init()
+class DataStorage:
+    __GROUP_MAP_FILE = "group.map"
+
+    def __init__(self, data_dir):
+        self.data_dir = data_dir
+
+    @staticmethod
+    def try_load_existed(data_dir):
+        pass
